requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
attrs==23.2.0
audioread==3.0.1
av==11.0.0
Babel==2.15.0
backoff==2.2.1
bangla==0.0.2
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
cairocffi==1.3.0
CairoSVG==2.7.1
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.12
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.2
confection==0.1.4
contourpy==1.2.1
coqpit==0.0.17
cssselect2==0.7.0
ctranslate2==4.2.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dataclasses-json==0.6.6
dateparser==1.1.8
decorator==5.1.1
defusedxml==0.7.1
diskcache==5.6.3
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
fastapi==0.99.1
faster-whisper==1.0.0
filelock==3.13.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.52.4
frozenlist==1.4.1
fsspec==2024.2.0
g2pkk==0.1.2
greenlet==3.0.3
grpcio==1.64.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.0
gruut_lang_en==2.0.0
gruut_lang_es==2.0.0
gruut_lang_fr==2.0.2
h11==0.14.0
hangul-romanize==0.1.0
httptools==0.6.1
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
importlib_resources==6.4.0
inflect==7.2.1
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.3
joblib==1.4.2
Js2Py==0.74
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==2.4
keyboard==0.13.5
kids.cache==0.0.7
kiwisolver==1.4.5
langchain==0.1.19
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.2
langcodes==3.4.0
langsmith==0.1.63
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
libsass==0.23.0
llama_cpp_python==0.2.73
llvmlite==0.42.0
lxml==5.2.2
marisa-trie==1.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
mdurl==0.1.2
mock==5.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.59.1
numpy==1.26.3
onnxruntime==1.18.0
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pillow==10.2.0
pipwin==0.5.2
platformdirs==4.2.2
pooch==1.8.1
posthog==3.5.0
preshed==3.0.9
protobuf==5.27.0
psutil==5.9.8
pulsar-client==3.5.0
pycparser==2.22
pydantic==1.10.15
pydantic_core==2.18.3
pygame==2.5.2
Pygments==2.18.0
pyjsparser==2.7.1
pynndescent==0.5.12
pyparsing==3.1.2
PyPika==0.48.9
pypinyin==0.51.0
pypiwin32==223
PyPrind==2.11.3
pyreadline3==3.4.1
pysbd==0.3.4
PySide6==6.6.3
PySide6_Addons==6.6.3
PySide6_Essentials==6.6.3
pySmartDL==1.3.4
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pyttsx3==2.90
pytz==2024.1
pywin32==306
PyYAML==6.0.1
QT-PyQt-PySide-Custom-Widgets==0.9.0
QtPy==2.4.1
qtsass==0.4.0
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.0
shellingham==1.5.4
shiboken6==6.6.3
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.0
SQLAlchemy==2.0.30
srsly==2.4.8
starlette==0.27.0
SudachiDict-core==20240409
SudachiPy==0.6.8
sympy==1.12
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.15.2
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu118
tqdm==4.66.1
trainer==0.0.36
transformers==4.38.1
TTS==0.22.0
typeguard==4.3.0
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.30.0
wasabi==1.1.2
watchdog==4.0.1
watchfiles==0.22.0
weasel==0.3.4
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.4
